+++ backend/tests/unit/api/Controllers/Note/NoteControllerTest.cs
@@ -83,7 +83,7 @@ namespace Pims.Api.Test.Controllers.Note
             var service = helper.GetService<Mock<INoteService>>();
             var mapper = helper.GetService<IMapper>();
 
-            service.Setup(m => m.DeleteNote(It.IsAny<NoteType>(), It.IsAny<int>()));
+            service.Setup(m => m.DeleteNote(It.IsAny<NoteType>(), It.IsAny<long>()));
 
             // Act
             var result = controller.DeleteNote(Constants.NoteType.Activity, 1);
@@ -91,7 +91,7 @@ namespace Pims.Api.Test.Controllers.Note
             // Assert
             var actionResult = Assert.IsType<JsonResult>(result);
             Assert.Equal(true, actionResult.Value);
-            service.Verify(m => m.DeleteNote(It.IsAny<NoteType>(), It.IsAny<int>()), Times.Once());
+            service.Verify(m => m.DeleteNote(It.IsAny<NoteType>(), It.IsAny<long>()), Times.Once());
         }
         #endregion
         #endregion
